Updated CSS styling, sanitized data inputs.

+++ signup.php
@@ -6,15 +6,18 @@ session_start();
     <head>
         <title>Pok&eacute;mon Type Modifier</title>
         <script type="text/javascript" src="login.js"></script>
+        <link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/css/bootstrap.min.css">
         <link rel="stylesheet" type="text/css" href="calculator.css">
     </head>
     <!--need php_readonly somewhere -->
     <body>
-        <div class="signup">      
+        <div class="login">      
         <h1><img src="pokemon.png" class="logo"/></h1>
-        <h2>Type Modifier Calculator</h2>
+        <h3>Type Modifier Calculator</h3>
 
         <?php
+
+        require "password.php";
         error_reporting(E_ALL);
         ini_set("display_errors", 1);
         $usernameSent = isset($_POST["username"]);
@@ -27,9 +30,11 @@ session_start();
         if (!$usernameSent || $username == "")
         {
             echo "<form id=\"signup\" action=\"signup.php\" method=\"POST\">";
-            echo "<p>Username <input type=\"input\" name=\"username\" id=\"username\"/></p>";
-            echo "<p>Password <input type=\"password\" name=\"password\" id=\"password\"/></p>";
-            echo "<button type=\"signup\" onclick=\"addUserData\">Sign Up</button>";
+            echo "<p>Username <input class=\"form-control\" type=\"input\" name=\"username\" id=\"username\"/></p>";
+            echo "<p>Password <input class=\"form-control\" type=\"password\" name=\"password\" id=\"password\"/></p>";
+            echo "<p>Confirm Password <input class=\"form-control\" type=\"password\" name=\"confirmPassword\" id=\"confirmPassword\"/></p>";
+            echo "</br>";
+            echo "<button class=\"btn btn-info btn-block\" type=\"signup\" onclick=\"addUserData\">Sign Up</button>";
             echo "</form>";
         }
         else
@@ -70,9 +75,22 @@ session_start();
             {
                 echo "<p>There is already a user by that name. Please choose a different username.</p>";
                 echo "<form id=\"signup\" action=\"signup.php\" method=\"POST\">";
-                echo "<p>Username <input type=\"input\" name=\"username\" id=\"username\"/></p>";
-                echo "<p>Password <input type=\"password\" name=\"password\" id=\"password\"/></p>";
-                echo "<button type=\"signup\">Sign Up</button>";
+                echo "<p>Username <input class=\"form-control\" type=\"input\" name=\"username\" id=\"username\"/></p>";
+                echo "<p>Password <input class=\"form-control\" type=\"password\" name=\"password\" id=\"password\"/></p>";
+                echo "<p>Confirm Password <input class=\"form-control\" type=\"password\" name=\"confirmPassword\" id=\"confirmPassword\"/></p>";
+                echo "</br>";
+                echo "<button class=\"btn btn-info btn-block\" type=\"signup\" onclick=\"addUserData\">Sign Up</button>";
+                echo "</form>";
+            }
+            else if($_POST['password'] != $_POST['confirmPassword'])
+            {
+                echo "<p>Your password's don't match. Please try again.</p>";
+                echo "<form id=\"signup\" action=\"signup.php\" method=\"POST\">";
+                echo "<p>Username <input class=\"form-control\" type=\"input\" name=\"username\" id=\"username\"/></p>";
+                echo "<p>Password <input class=\"form-control\" type=\"password\" name=\"password\" id=\"password\"/></p>";
+                echo "<p>Confirm Password <input class=\"form-control\" type=\"password\" name=\"confirmPassword\" id=\"confirmPassword\"/></p>";
+                echo "</br>";
+                echo "<button class=\"btn btn-info btn-block\" type=\"signup\" onclick=\"addUserData\">Sign Up</button>";
                 echo "</form>";
             }
             else
@@ -81,6 +99,7 @@ session_start();
                 $password = $_POST['password'];
                 $qry = $db->prepare("INSERT INTO users (username, password) VALUES (:username, :password);");
                 $qry->bindValue(':username', $username);
+                $password = password_hash($password, PASSWORD_DEFAULT);
                 $qry->bindValue(':password', $password);
                 $qry->execute();
